changed serial to bigserial and integer to bigint

+++ migrations/2022-02-04-013022_create_assignments/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE assignments (
-    ticket_id INTEGER REFERENCES tickets (id) NOT NULL,
-    user_id INTEGER REFERENCES users (id) NOT NULL,
+    ticket_id BIGINT REFERENCES tickets (id) NOT NULL,
+    user_id BIGINT REFERENCES users (id) NOT NULL,
     assigned TIMESTAMP NOT NULL,
     PRIMARY KEY (ticket_id, user_id)
 );
