+++ migrations/2022-02-04-020229_create_comments/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE comments (
-    id SERIAL PRIMARY KEY,
-    ticket_id INTEGER REFERENCES tickets (id) NOT NULL,
-    commenter_id INTEGER REFERENCES users (id) NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    ticket_id BIGINT REFERENCES tickets (id) NOT NULL,
+    commenter_id BIGINT REFERENCES users (id) NOT NULL,
     content TEXT NOT NULL,
     commented TIMESTAMP NOT NULL
 );
